+++ src/index.js
@@ -1,7 +1,7 @@
 /**
  * @license MIT
  * @name tracking-manager
- * @version 1.0.1
+ * @version 2.0.0
  * @author: Yoriiis aka Joris DANIEL <joris.daniel@gmail.com>
  * @description: Tracking Manager allows to manage all your Google Analytics events directly in HTML or Javascript with a simple and extensible configuration and public functions to track events with dynamic variables
  * {@link https://github.com/yoriiis/tracking-manager}
@@ -10,37 +10,48 @@
 
 module.exports = class Tracking {
 	/**
-	 * Instanciate the constructor
 	 *
-	 * @param {Object} config Tracking configurations
-	 * @param {Boolean} debug Debug mode
+	 * @param {Object} options Options parameters
+	 * @param {Object} options.config Tracking config
+	 * @param {Object} options.type Tracking type (GA|GTM)
 	 */
-	constructor ({ config, debug = false }) {
-		this.config = config;
-		this.debug = debug;
+	constructor({ config, type = 'ga' }) {
 		this.selector = '[data-track]';
+		this.config = config;
+		this.type = type;
 		this.ignoreRedirectAttribute = 'data-no-tracking-redirect';
-	}
-
-	/**
-	 * Instanciate the Tracking class
-	 */
-	init () {
-		this.parseDOM();
+		this.trackClickEvent = this.trackClickEvent.bind(this);
 	}
 
 	/**
 	 * Parse the DOM to search all tracking attributes
+	 *
+	 * @param {HTMLElement} domElement Target element to search "data-track" attributes
 	 */
-	parseDOM () {
-		const elementsToTrack = [
-			...document.querySelectorAll(`${this.selector}:not([tracking-parsed])`)
-		];
+	parseDom(domElement) {
+		if (this.isTrackingAvailable(domElement)) {
+			const elementsToTrack = [
+				...domElement.querySelectorAll(`${this.selector}:not([tracking-parsed])`)
+			];
+			elementsToTrack.forEach((element) => {
+				element.setAttribute('tracking-parsed', '');
+				element.addEventListener('click', this.trackClickEvent);
+			});
+		}
+	}
 
-		elementsToTrack.forEach(element => {
-			element.setAttribute('tracking-parsed', '');
-			element.addEventListener('click', this.trackClickEvent.bind(this), false);
-		});
+	/**
+	 * Check if the tracking can be send
+	 *
+	 * @param {HTMLElement} domElement Target element to search "data-track" attributes
+	 *
+	 * @returns {Boolean} Tracking status
+	 */
+	isTrackingAvailable(domElement) {
+		return (
+			domElement &&
+			((this.type === 'ga' && this.isGoogleAnalyticsAvailable()) || this.type === 'gtm')
+		);
 	}
 
 	/**
@@ -49,7 +60,7 @@ module.exports = class Tracking {
 	 * @param {String} key Key of tracking datas
 	 * @param {Object} replaceObj JSON with replacement keys
 	 */
-	trackPageView (key, pageView) {
+	trackPageView(key, pageView) {
 		let dataTracking = {};
 		const configEvent = this.getConfigEventFromKey(key);
 
@@ -72,7 +83,7 @@ module.exports = class Tracking {
 	 * @param {String} key Key of tracking datas
 	 * @param {Object} replaceObj JSON with replacement keys
 	 */
-	trackEvent (key, replaceObj) {
+	trackEvent(key, replaceObj) {
 		let dataTracking = {};
 		const configEvent = this.getConfigEventFromKey(key);
 
@@ -84,9 +95,7 @@ module.exports = class Tracking {
 
 		this.sendEvent({
 			key: key,
-			json: dataTracking,
-			callbackUrl: false,
-			target: false
+			json: dataTracking
 		});
 	}
 
@@ -95,17 +104,17 @@ module.exports = class Tracking {
 	 *
 	 * @param {Object} e Event listener datas
 	 */
-	trackClickEvent (e) {
-		const link = e.currentTarget;
-		const key = link.getAttribute('data-track-key');
-		const dataTrackParams = link.getAttribute('data-track-params');
-		const isPageView = link.hasAttribute('data-track-page-view');
+	trackClickEvent(e) {
+		const element = e.currentTarget;
+		const key = element.getAttribute('data-track-key');
+		const dataTrackParams = element.getAttribute('data-track-params');
+		const isPageView = element.hasAttribute('data-track-page-view');
 		let dataTracking = {};
-		const href = link.getAttribute('href') || false;
-		const target = link.getAttribute('target') || false;
+		const callbackUrl = element.getAttribute('href') || false;
+		const targetAttribute = element.getAttribute('target') || false;
 
 		// Prevent only when link has no target attribute
-		if (target === false) {
+		if (!targetAttribute) {
 			e.preventDefault();
 		}
 
@@ -114,7 +123,7 @@ module.exports = class Tracking {
 
 		// Check if element contain page view attribute
 		if (isPageView) {
-			const pageView = link.getAttribute('data-track-page-view');
+			const pageView = element.getAttribute('data-track-page-view');
 
 			configEvent.pageView = pageView;
 			dataTracking = configEvent;
@@ -133,11 +142,11 @@ module.exports = class Tracking {
 
 			// Send event
 			this.sendEvent({
-				key: key,
+				key,
 				json: dataTracking,
-				callbackUrl: href,
-				target: target,
-				element: link
+				callbackUrl,
+				targetAttribute,
+				element
 			});
 		}
 	}
@@ -150,7 +159,7 @@ module.exports = class Tracking {
 	 *
 	 * @returns {Object} Object with all replacements
 	 */
-	loopReplace (obj, replaceObj) {
+	loopReplace(obj, replaceObj) {
 		let replacedObj = {};
 
 		if (!replaceObj) {
@@ -172,39 +181,35 @@ module.exports = class Tracking {
 	/**
 	 * Send Google Analytics event with all parameters
 	 *
-	 * @param {String} key Key of tracking datas
-	 * @param {Object} json Object to send to GA
-	 * @param {String} callbackUrl Url to callback if preventDefault has been executed
-	 * @param {Boolean} target Is target attribute is present on the element
-	 * @param {Object} element Source element of the event
+	 * @param {Object} options Options parameters
+	 * @param {String} options.key Key of tracking datas
+	 * @param {Object} options.json Object to send to GA
+	 * @param {(String|Boolean)} options.callbackUrl Url to redirect if necessary
+	 * @param {Boolean} options.targetAttribute Is target attribute is present on the element
+	 * @param {(Object|Boolean)} options.element Source element of the event
 	 */
-	sendEvent ({ key, json, callbackUrl, target, element = null }) {
-		if (this.debug) {
-			console.log('%c[Tracking -> trackEvent]:', 'color: DeepPink;', key, json, {
-				callbackUrl: callbackUrl,
-				target: target,
-				element: element
-			});
-		}
-
-		// Possibility to disable the redirect in case of JS events conflicts (with PMC for example)
-		const ignoreRedirect =
-			element !== null && element.hasAttribute(this.ignoreRedirectAttribute);
-
-		if (this.isGoogleAnalyticsAvailable()) {
-			// Redirect is necessary on event callback
-			if (
-				callbackUrl &&
-				callbackUrl !== '' &&
-				callbackUrl.indexOf('#') !== 0 &&
-				!target &&
-				!ignoreRedirect
-			) {
+	sendEvent({ key, json = {}, callbackUrl = false, targetAttribute = false, element = false }) {
+		console.log('%c[Tracking -> trackEvent]:', 'color: DeepPink;', key, json, {
+			callbackUrl,
+			targetAttribute,
+			element
+		});
+		const needRedirect = this.needRedirectAfterEvent({
+			element,
+			callbackUrl,
+			targetAttribute
+		});
+		if (this.type === 'gtm') {
+			window.dataLayer.push(json);
+			if (needRedirect) {
+				window.location.assign(callbackUrl);
+			}
+		} else {
+			if (needRedirect) {
 				json.hitCallback = () => {
 					window.location.assign(callbackUrl);
 				};
 			}
-
 			window.ga('send', json);
 		}
 	}
@@ -212,17 +217,26 @@ module.exports = class Tracking {
 	/**
 	 * Send Google Analytics page view with all parameters
 	 *
-	 * @param {String} key Key of tracking datas
-	 * @param {Object} json Object to send to GA
+	 * @param {Object} options Options parameters
+	 * @param {String} options.key Key of tracking datas
+	 * @param {Object} options.json Object to send to GA
 	 */
-	sendPageView ({ key, json }) {
-		if (this.debug) {
-			console.log('%c[Tracking -> trackPageView]:', 'color: DeepPink;', key, json);
-		}
+	sendPageView({ key, json }) {
+		console.log('%c[Tracking -> trackPageView]:', 'color: DeepPink;', key, json);
 
-		if (this.isGoogleAnalyticsAvailable() && json.pageView) {
-			window.ga('set', 'page', json.pageView);
-			window.ga('send', 'pageView');
+		if (json.pageView) {
+			if (this.type === 'gtm') {
+				window.dataLayer.push({
+					event: 'pageview',
+					page: {
+						path: json.pageView,
+						title: document.title
+					}
+				});
+			} else {
+				window.ga('set', 'page', json.pageView);
+				window.ga('send', 'pageView');
+			}
 		}
 	}
 
@@ -231,7 +245,7 @@ module.exports = class Tracking {
 	 *
 	 * @return {Boolean} Is Google Analytics available
 	 */
-	isGoogleAnalyticsAvailable () {
+	isGoogleAnalyticsAvailable() {
 		return typeof window.ga !== 'undefined';
 	}
 
@@ -243,9 +257,27 @@ module.exports = class Tracking {
 	 *
 	 * @return {Object} Tracking configuration datas
 	 */
-	getConfigEventFromKey (key) {
+	getConfigEventFromKey(key) {
 		return key
 			.split('.')
 			.reduce((accumulator, currentValue) => accumulator[currentValue], this.config);
+	}
+
+	/**
+	 * Check if redirect is needed after the event
+	 *
+	 * @param {Object} options Options parameters
+	 * @param {HTMLElement} options.callbackUrl The HTML element which trigger the event
+	 * @param {HTMLElement} options.targetAttribute The HTML element which trigger the event
+	 * @param {HTMLElement} options.element The HTML element which trigger the event
+	 */
+	needRedirectAfterEvent({ callbackUrl, targetAttribute, element }) {
+		return !!(
+			callbackUrl &&
+			callbackUrl !== '' &&
+			callbackUrl !== '#' &&
+			!targetAttribute &&
+			!element.hasAttribute(this.ignoreRedirectAttribute)
+		);
 	}
 };