+++ CHANGELOG.md
@@ -1,14 +1,26 @@
 # CHANGELOG
 
+## 2.0.0
+
+### New features
+
+- Add the Google Analytics and Google Tag Manager compatibility (`type` parameter)
+
+### Updates
+
+- Move CI to GitHub Actions
+- Update th node.js engine to `v12.14.0`
+- Move the tracking configuration on each instance (tracking by components)
+
 ## 1.0.1
 
 ### Updates
 
-* Update Jest tests
+- Update Jest tests
 
 ## 1.0.0
 
 ### New features
 
-* First release of `tracking-manager` 🚀
-* Jest coverage of 100%
+- First release of `tracking-manager` 🚀
+- Jest coverage of 100%